pyproject.toml
[project]
name = "pydantic-ai-mem0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mem0ai>=0.1.90",
    "pydantic-ai>=0.0.55",
    "pydantic-ai-slim>=0.0.55",
    "python-dotenv>=1.1.0",
]
